expose writer and metadata
