+++ rust/lance-file/Cargo.toml
@@ -37,7 +37,6 @@ tracing.workspace = true
 rand.workspace = true
 tempfile.workspace = true
 proptest.workspace = true
-lance-table.workspace = true
 
 [build-dependencies]
 prost-build.workspace = true
