+++ rust/lance-file/src/writer.rs
@@ -86,8 +86,8 @@ impl ManifestProvider for NotSelfDescribing {
 /// file_writer.shutdown();
 /// ```
 pub struct FileWriter<M: ManifestProvider + Send + Sync> {
-    object_writer: ObjectWriter,
-    schema: Schema,
+    pub object_writer: ObjectWriter,
+    pub schema: Schema,
     batch_id: i32,
     page_table: PageTable,
     metadata: Metadata,
@@ -719,7 +719,6 @@ mod tests {
         DataType, Field as ArrowField, Fields as ArrowFields, Schema as ArrowSchema, TimeUnit,
     };
     use arrow_select::concat::concat_batches;
-    use lance_table::io::manifest::ManifestDescribing;
     use object_store::path::Path;
 
     use crate::reader::FileReader;
@@ -1117,7 +1116,7 @@ mod tests {
                 min_values: Arc::new(Int64Array::from(vec![1, 4])),
                 max_values: Arc::new(Int64Array::from(vec![3, 5])),
             },
-            // FIXME: these max values shouldn't be incremented
+            // 2FIXME: these max values shouldn't be incremented
             // https://github.com/lancedb/lance/issues/1517
             Stats {
                 field_id: 6,
@@ -1242,42 +1241,5 @@ mod tests {
 
         let batch = read_file_as_one_batch(&store, &path, schema).await;
         assert_eq!(batch.column_by_name("i").unwrap().as_ref(), &array);
-    }
-
-    #[tokio::test]
-    async fn test_update_schema_metadata() {
-        let store = ObjectStore::memory();
-        let path = Path::from("/update_schema_metadata");
-
-        let arrow_schema = Arc::new(ArrowSchema::new(vec![ArrowField::new(
-            "i",
-            DataType::Int32,
-            false,
-        )]));
-        let schema = Schema::try_from(arrow_schema.as_ref()).unwrap();
-        let mut file_writer = FileWriter::<ManifestDescribing>::try_new(
-            &store,
-            &path,
-            schema.clone(),
-            &Default::default(),
-        )
-        .await
-        .unwrap();
-
-        let array = Int32Array::from_iter_values(0..10);
-        let batch =
-            RecordBatch::try_new(arrow_schema.clone(), vec![Arc::new(array.clone())]).unwrap();
-        file_writer.write(&[batch.clone()]).await.unwrap();
-        file_writer
-            .schema
-            .metadata
-            .insert(String::from("lance:extra"), String::from("for_test"));
-        file_writer.finish().await.unwrap();
-
-        let reader = FileReader::try_new(&store, &path, schema.clone())
-            .await
-            .unwrap();
-        let schema = ArrowSchema::from(reader.schema());
-        assert_eq!(schema.metadata().get("lance:extra").unwrap(), "for_test");
     }
 }
